Update version field type from Long to BigInteger in enforcement and account-related tests

+++ src/test/java/uk/gov/hmcts/opal/service/DefendantAccountEnforcementServiceTest.java
@@ -64,7 +64,7 @@ class DefendantAccountEnforcementServiceTest {
         AddEnforcementResponse proxyResponse = AddEnforcementResponse.builder()
             .enforcementId("ENF123")
             .defendantAccountId("77")
-            .version(3)
+            .version(BigInteger.valueOf(3))
             .build();
 
         when(userStateService.getUserStateV1FromSecurityContext()).thenReturn(userState);
